Adds const uint8 cClient::buttons to AS
This property allows gametypes to check the pressed buttons of a given
client. This way a gametype can detect whether a player is pressing his
attack button at a given time. This could be used to to implement
controllable vehicles for example. Intercepting the +attack command has
proven to be an unreliable method to accomplish something like that.