Reintroduces ServerCommand "cmd"
This command was removed in commit c1fb9bcaf17b721badb38e410c8f3c4ac4e79527
while reworking the way 'Move to TV' works. That function was the only
one calling 'cmd' in the engine, so removing it seemed to be fine.
However GameCommands are exported to gametype AS via
Client::execGameCommand( String ), so removing the command broke
gametype scripts.